UI Bugs fixed
